# sources/antimage-dev/internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "antimage API",
    "version": "0.3.0",
    "description": "Go-native antimage panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `antimage_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current antimage panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/node": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Create node",
        "description": "Create a node using generated or pending certificate material and enqueue initial sync.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Requires sudo/full_access. Node calls use Go node controller and gRPC/mTLS where runtime communication is needed.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Create payload may include certificate material directly or a `certificate_token` returned by `/api/node/certificate/new`. The response includes `node_certificate` and one-time `node_certificate_key`; the dashboard combines them into a single node install bundle for the installer. `add_as_new_host` is removed and ignored by the Go flow.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postNode"
      }
    },
...
    "/api/inbounds": {
      "get": {
        "tags": [
          "Inbounds"
        ],
        "summary": "List inbounds",
        "description": "Return manageable inbounds grouped by protocol.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Listing is readable by authenticated admins; create/update/delete require sudo/full_access.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getInbounds"
      },
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Create inbound",
        "description": "Create a manageable inbound in the selected config target.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Listing is readable by authenticated admins; create/update/delete require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postInbounds"
      }
    },
...
    "/inbounds": {
      "get": {
        "tags": [
          "Inbounds"
        ],
        "summary": "List inbounds alias",
        "description": "Compatibility alias for /api/inbounds.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Listing is readable by authenticated admins; create/update/delete require sudo/full_access.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getInbounds"
      },
      "post": {
        "tags": [
          "Inbounds"
        ],
        "summary": "Create inbound alias",
        "description": "Compatibility alias for /api/inbounds.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Listing is readable by authenticated admins; create/update/delete require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postInbounds"
      }
    },
...
    "/api/settings/subscriptions": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get subscription settings",
        "description": "Return global subscription settings, templates metadata, aliases, ports, and admin overrides.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getSettingsSubscriptions"
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update subscription settings",
        "description": "Update global subscription settings, aliases, ports, template directories, and client format flags.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putSettingsSubscriptions"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "antimage admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "antimage_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            },
            "example": {
              "name": "example",
              "status": "active",
              "settings": {}
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible antimage API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "name": "example",
          "status": "active",
          "settings": {}
        }
      },
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}